@@ -1,43 +1,48 @@
 {
-  "name": "slush-node-typescript",
-  "description": "A slush generator for Node.js and TypeScript.",
-  "version": "0.4.3",
-  "homepage": "https://github.com/shiwano/slush-node-typescript",
+  "name": "slush-node-typescript2",
+  "description": "A slush generator for Node.js +6 and TypeScript 2.x with: Gulp/Webpack (task managers), Mocha & Chai (unit tests), Istanbul (test coverage) and Nightwatch (E2E testing), and Sinon (server mocking)",
+  "version": "0.5.0",
+  "homepage": "https://github.com/kristianmandrup/slush-node-typescript2",
   "author": {
-    "name": "Shogo Iwano",
-    "email": "shiwano@gmail.com"
+    "name": "Kristian Mandrup",
+    "email": "kmandrup@gmail.com"
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/shiwano/slush-node-typescript.git"
+    "url": "git://github.com/kristianmandrup/slush-node-typescript2.git"
   },
   "bugs": {
-    "url": "https://github.com/shiwano/slush-node-typescript/issues"
+    "url": "https://github.com/kristianmandrup/slush-node-typescript2/issues"
   },
   "licenses": [
     {
       "type": "MIT",
-      "url": "https://github.com/shiwano/slush-node-typescript/blob/master/LICENSE"
+      "url": "https://github.com/kristianmandrup/slush-node-typescript2/blob/master/LICENSE"
     }
   ],
   "main": "slushfile.js",
   "engines": {
-    "node": ">= 0.10.0",
-    "npm": ">=1.4.0"
+    "node": ">= 6.0.0",
+    "npm": ">=3.0.0"
   },
   "dependencies": {
     "git-config": "0.0.6",
-    "gulp": "^3.8.5",
-    "gulp-conflict": "^0.1.1",
-    "gulp-install": "^0.1.5",
-    "gulp-rename": "^1.2.0",
-    "gulp-template": "^0.1.1",
-    "inflection": "^1.5.0",
-    "inquirer": "^0.5.0",
-    "license": "0.0.1"
+    "gulp": "^3.9.1",
+    "gulp-conflict": "^0.4.0",
+    "gulp-install": "^0.6.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-ignore": "^2.0.1",
+    "gulp-template": "^4.0.0",
+    "inflection": "^1.10.0",
+    "inquirer": "^1.1.3",
+    "license": "kristianmandrup/license#master"
   },
   "keywords": [
     "slushgenerator",
-    "typescript"
+    "node",
+    "typescript",
+    "mocha",
+    "gulp",
+    "webpack"
   ]
 }
